refactor: extract remaining search datafusion optimizers
