+++ src/core/src/service/search/index.rs
@@ -5,7 +5,7 @@
 // the Free Software Foundation, either version 3 of the License, or
 // (at your option) any later version.
 //
-// This program is distributed in the hope that it will be useful
+// This program is distributed in the hope that it will be useful,
 // but WITHOUT ANY WARRANTY; without even the implied warranty of
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 // GNU Affero General Public License for more details.
@@ -13,1456 +13,4 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-use std::{
-    fmt::{self, Debug, Formatter},
-    sync::Arc,
-};
-
-use config::{
-    INDEX_FIELD_NAME_FOR_ALL, get_config,
-    meta::inverted_index::UNKNOWN_NAME,
-    tantivy::{query::contains_query::ContainsQuery, tokenizer::o2_collect_search_tokens},
-};
-use datafusion::{
-    arrow::datatypes::{DataType, SchemaRef},
-    config::ConfigOptions,
-    logical_expr::Operator,
-    physical_expr::{ScalarFunctionExpr, conjunction},
-    physical_plan::{
-        PhysicalExpr,
-        expressions::{
-            BinaryExpr, CastExpr, Column, InListExpr, IsNotNullExpr, LikeExpr, Literal, NotExpr,
-        },
-    },
-    scalar::ScalarValue,
-};
-use hashbrown::HashSet;
-use tantivy::{
-    Term,
-    query::{
-        AllQuery, BooleanQuery, FuzzyTermQuery, Occur, PhrasePrefixQuery, Query, RegexQuery,
-        TermQuery,
-    },
-    schema::{Field, IndexRecordOption, Schema},
-};
-
-use super::datafusion::udf::fuzzy_match_udf;
-use crate::service::search::datafusion::udf::{
-    MATCH_FIELD_IGNORE_CASE_UDF_NAME, MATCH_FIELD_UDF_NAME, STR_MATCH_UDF_IGNORE_CASE_NAME,
-    STR_MATCH_UDF_NAME,
-    match_all_udf::{FUZZY_MATCH_ALL_UDF_NAME, MATCH_ALL_UDF_NAME},
-    str_match_udf,
-};
-
-// note the condition in IndexCondition is connection by AND operator
-#[derive(Default, Clone, Hash, Eq, PartialEq)]
-pub struct IndexCondition {
-    pub conditions: Vec<Condition>,
-}
-
-impl IndexCondition {
-    pub fn new() -> Self {
-        IndexCondition {
-            conditions: Vec::new(),
-        }
-    }
-
-    pub fn is_empty(&self) -> bool {
-        self.conditions.is_empty()
-    }
-
-    pub fn add_condition(&mut self, condition: Condition) {
-        self.conditions.push(condition);
-    }
-}
-
-impl Debug for IndexCondition {
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.to_query())
-    }
-}
-
-impl IndexCondition {
-    // this only use for display the query
-    pub fn to_query(&self) -> String {
-        self.conditions
-            .iter()
-            .map(|condition| condition.to_query())
-            .collect::<Vec<_>>()
-            .join(" AND ")
-    }
-
-    // get the tantivy query for the index condition
-    // Returns (query, has_skipped):
-    //   has_skipped = true means some conditions were skipped because the field
-    //   does not exist in this tantivy index (e.g., a newly added index field
-    //   that has no historical data). The caller must keep the DataFusion filter
-    //   so that the skipped predicates are still evaluated.
-    pub fn to_tantivy_query(
-        &self,
-        trace_id: &str,
-        schema: Schema,
-        default_field: Option<Field>,
-    ) -> anyhow::Result<(Box<dyn Query>, bool)> {
-        let mut has_skipped = false;
-        let mut queries: Vec<(Occur, Box<dyn Query>)> = Vec::with_capacity(self.conditions.len());
-        for condition in &self.conditions {
-            match condition.to_tantivy_query(&schema, default_field) {
-                Ok(query) => {
-                    queries.push((Occur::Must, query));
-                }
-                Err(e) => {
-                    log::info!(
-                        "[trace_id {trace_id}] to_tantivy_query: skipping condition due to error: {e}"
-                    );
-                    has_skipped = true;
-                }
-            }
-        }
-        if queries.is_empty() {
-            Err(anyhow::anyhow!(
-                "All AND conditions are failed to generate tantivy query"
-            ))
-        } else if queries.len() == 1 {
-            Ok((queries.pop().unwrap().1, has_skipped))
-        } else {
-            Ok((Box::new(BooleanQuery::from(queries)), has_skipped))
-        }
-    }
-
-    // get the fields use for search in datafusion(for add filter back logical)
-    pub fn get_schema_fields(&self, fst_fields: &[String]) -> HashSet<String> {
-        self.conditions
-            .iter()
-            .fold(HashSet::new(), |mut acc, condition| {
-                acc.extend(condition.get_schema_fields(fst_fields));
-                acc
-            })
-    }
-
-    pub fn get_schema_projection(&self, schema: SchemaRef, fst_fields: &[String]) -> Vec<usize> {
-        let fields = self.get_schema_fields(fst_fields);
-        let mut projection = Vec::with_capacity(fields.len());
-        for field in fields.iter() {
-            if let Ok(index) = schema.index_of(field) {
-                projection.push(index);
-            }
-        }
-        projection
-    }
-
-    pub fn need_all_term_fields(&self) -> Vec<String> {
-        self.conditions
-            .iter()
-            .flat_map(|condition| condition.need_all_term_fields())
-            .collect::<HashSet<_>>()
-            .into_iter()
-            .collect()
-    }
-
-    pub fn to_physical_expr(
-        &self,
-        schema: &arrow_schema::Schema,
-        fst_fields: &[String],
-    ) -> Result<Arc<dyn PhysicalExpr>, anyhow::Error> {
-        Ok(conjunction(
-            self.conditions
-                .iter()
-                .map(|condition| condition.to_physical_expr(schema, fst_fields))
-                .collect::<Result<Vec<_>, _>>()?,
-        ))
-    }
-
-    pub fn can_remove_filter(&self) -> bool {
-        self.conditions
-            .iter()
-            .all(|condition| condition.can_remove_filter())
-    }
-
-    // use for simple distinct optimization
-    pub fn get_str_match_condition(&self) -> Option<(String, bool)> {
-        match &self.conditions[0] {
-            Condition::StrMatch(_, value, case_sensitive) => {
-                Some((value.to_string(), *case_sensitive))
-            }
-            Condition::All() => None, // for the condition that query without filter
-            _ => unreachable!("get_str_match_condition only support one str_match condition"),
-        }
-    }
-
-    // use for check if the index condition is only
-    // for the condition that query without filter
-    pub fn is_condition_all(&self) -> bool {
-        self.conditions.len() == 1 && matches!(self.conditions[0], Condition::All())
-    }
-
-    // use for the simple histogram RANK fast path: the single `field = value` term
-    pub fn single_equal_term(&self) -> Option<(String, String)> {
-        if self.conditions.len() == 1
-            && let Condition::Equal(field, value) = &self.conditions[0]
-        {
-            Some((field.clone(), value.clone()))
-        } else {
-            None
-        }
-    }
-}
-
-// single condition
-#[derive(Debug, Clone, Hash, Eq, PartialEq)]
-pub enum Condition {
-    // field, value
-    Equal(String, String),
-    // field, value
-    NotEqual(String, String),
-    // field, value, case_sensitive
-    StrMatch(String, String, bool),
-    // field, values, negated
-    In(String, Vec<String>, bool),
-    // field, pattern
-    Regex(String, String),
-    // term
-    MatchAll(String),
-    // term, distance
-    FuzzyMatchAll(String, u8),
-    All(),
-    Or(Box<Condition>, Box<Condition>),
-    And(Box<Condition>, Box<Condition>),
-    Not(Box<Condition>),
-}
-
-impl Condition {
-    // this only use for display the query
-    pub fn to_query(&self) -> String {
-        match self {
-            Condition::Equal(field, value) => format!("{field}={value}"),
-            Condition::NotEqual(field, value) => format!("{field}!={value}"),
-            Condition::StrMatch(field, value, case_sensitive) => {
-                if *case_sensitive {
-                    format!("str_match({field}, '{value}')")
-                } else {
-                    format!("str_match_ignore_case({field}, '{value}')")
-                }
-            }
-            Condition::In(field, values, negated) => {
-                if *negated {
-                    format!("{} NOT IN ({})", field, values.join(","))
-                } else {
-                    format!("{} IN ({})", field, values.join(","))
-                }
-            }
-            Condition::Regex(field, value) => format!("{field}=~{value}"),
-            Condition::MatchAll(value) => {
-                let tokens = o2_collect_search_tokens(value);
-                format!("({INDEX_FIELD_NAME_FOR_ALL}:{value}):({tokens:?})")
-            }
-            Condition::FuzzyMatchAll(value, distance) => {
-                format!("{INDEX_FIELD_NAME_FOR_ALL}:fuzzy({value}, {distance})")
-            }
-            Condition::All() => "ALL".to_string(),
-            Condition::Or(left, right) => format!("({} OR {})", left.to_query(), right.to_query()),
-            Condition::And(left, right) => {
-                format!("({} AND {})", left.to_query(), right.to_query())
-            }
-            Condition::Not(condition) => format!("NOT({})", condition.to_query()),
-        }
-    }
-
-    pub fn from_physical_expr(expr: &Arc<dyn PhysicalExpr>) -> Self {
-        if let Some(expr) = expr.downcast_ref::<BinaryExpr>() {
-            match expr.op() {
-                Operator::Eq | Operator::NotEq => {
-                    let (field, value) = if is_physical_value(expr.left())
-                        && is_physical_column(expr.right())
-                    {
-                        (
-                            get_physical_column_name(expr.right()).to_string(),
-                            get_physical_value(expr.left()),
-                        )
-                    } else if is_physical_value(expr.right()) && is_physical_column(expr.left()) {
-                        (
-                            get_physical_column_name(expr.left()).to_string(),
-                            get_physical_value(expr.right()),
-                        )
-                    } else {
-                        unreachable!()
-                    };
-
-                    if *expr.op() == Operator::Eq {
-                        Condition::Equal(field, value)
-                    } else {
-                        Condition::NotEqual(field, value)
-                    }
-                }
-                Operator::And => Condition::And(
-                    Box::new(Condition::from_physical_expr(expr.left())),
-                    Box::new(Condition::from_physical_expr(expr.right())),
-                ),
-                Operator::Or => Condition::Or(
-                    Box::new(Condition::from_physical_expr(expr.left())),
-                    Box::new(Condition::from_physical_expr(expr.right())),
-                ),
-                _ => unreachable!(),
-            }
-        } else if let Some(expr) = expr.downcast_ref::<InListExpr>() {
-            let field = get_physical_column_name(expr.expr()).to_string();
-            let values = expr.list().iter().map(get_physical_value).collect();
-            Condition::In(field, values, expr.negated())
-        } else if let Some(expr) = expr.downcast_ref::<ScalarFunctionExpr>() {
-            let name = expr.name();
-            match name {
-                MATCH_ALL_UDF_NAME => Condition::MatchAll(get_physical_value(&expr.args()[0])),
-                FUZZY_MATCH_ALL_UDF_NAME => {
-                    let value = get_physical_value(&expr.args()[0]);
-                    let distance = get_physical_value(&expr.args()[1]).parse().unwrap_or(1);
-                    Condition::FuzzyMatchAll(value, distance)
-                }
-                STR_MATCH_UDF_NAME | MATCH_FIELD_UDF_NAME => {
-                    let field = get_physical_column_name(&expr.args()[0]).to_string();
-                    let value = get_physical_value(&expr.args()[1]);
-                    Condition::StrMatch(field, value, true)
-                }
-                STR_MATCH_UDF_IGNORE_CASE_NAME | MATCH_FIELD_IGNORE_CASE_UDF_NAME => {
-                    let field = get_physical_column_name(&expr.args()[0]).to_string();
-                    let value = get_physical_value(&expr.args()[1]);
-                    Condition::StrMatch(field, value, false)
-                }
-                _ => unreachable!(),
-            }
-        } else if let Some(expr) = expr.downcast_ref::<NotExpr>() {
-            Condition::Not(Box::new(Condition::from_physical_expr(expr.arg())))
-        } else {
-            unreachable!()
-        }
-    }
-
-    pub fn to_tantivy_query(
-        &self,
-        schema: &Schema,
-        default_field: Option<Field>,
-    ) -> anyhow::Result<Box<dyn Query>> {
-        Ok(match self {
-            Condition::Equal(field, value) => {
-                let field = schema.get_field(field)?;
-                let term = Term::from_field_text(field, value);
-                Box::new(TermQuery::new(term, IndexRecordOption::Basic))
-            }
-            Condition::NotEqual(field, value) => {
-                let field = schema.get_field(field)?;
-                let term = Term::from_field_text(field, value);
-                let query = Box::new(TermQuery::new(term, IndexRecordOption::Basic));
-                Box::new(BooleanQuery::new(vec![
-                    (Occur::MustNot, query),
-                    (Occur::Must, Box::new(AllQuery {})),
-                ]))
-            }
-            Condition::In(field, values, negated) => {
-                let field = schema.get_field(field)?;
-                let terms: Vec<Box<dyn Query>> = values
-                    .iter()
-                    .map(|value| {
-                        let term = Term::from_field_text(field, value);
-                        Box::new(TermQuery::new(term, IndexRecordOption::Basic)) as _
-                    })
-                    .collect();
-                let query = Box::new(BooleanQuery::union(terms));
-                if *negated {
-                    Box::new(BooleanQuery::new(vec![
-                        (Occur::MustNot, query),
-                        (Occur::Must, Box::new(AllQuery {})),
-                    ]))
-                } else {
-                    query
-                }
-            }
-            Condition::Regex(field, value) => {
-                let field = schema.get_field(field)?;
-                Box::new(RegexQuery::from_pattern(value, field)?)
-            }
-            Condition::StrMatch(field, value, case_sensitive) => {
-                let field = schema.get_field(field)?;
-                Box::new(ContainsQuery::new(value, field, *case_sensitive)?)
-            }
-            Condition::MatchAll(value) => {
-                let default_field = default_field.ok_or_else(|| {
-                    anyhow::anyhow!("There's no FullTextSearch field for match_all() function")
-                })?;
-                if value.is_empty() || value == "*" {
-                    Box::new(AllQuery {})
-                } else {
-                    let mut tokens = o2_collect_search_tokens(value);
-                    let contains_search =
-                        tokens.len() == 1 && value.starts_with("*") && value.ends_with("*");
-                    let first_prefix = if value.starts_with("*") && !tokens.is_empty() {
-                        Some(tokens.remove(0))
-                    } else {
-                        None
-                    };
-                    let last_prefix = if value.ends_with("*") {
-                        tokens.pop()
-                    } else {
-                        None
-                    };
-                    let mut terms: Vec<Box<dyn Query>> = tokens
-                        .into_iter()
-                        .map(|value| {
-                            let term = Term::from_field_text(default_field, &value);
-                            Box::new(TermQuery::new(term, IndexRecordOption::Basic)) as _
-                        })
-                        .collect();
-                    if let Some(value) = first_prefix {
-                        terms.push(if contains_search {
-                            Box::new(ContainsQuery::new_case_insensitive(&value, default_field)?)
-                        } else {
-                            let value = format!(".*{value}");
-                            Box::new(RegexQuery::from_pattern(&value, default_field)?)
-                        });
-                    }
-                    if let Some(value) = last_prefix {
-                        terms.push(Box::new(PhrasePrefixQuery::new_with_offset(vec![(
-                            0,
-                            Term::from_field_text(default_field, &value),
-                        )])));
-                    }
-                    if !terms.is_empty() {
-                        if terms.len() > 1 {
-                            Box::new(BooleanQuery::intersection(terms))
-                        } else {
-                            terms.remove(0)
-                        }
-                    } else {
-                        return Err(anyhow::anyhow!(
-                            "The value of match_all() function can't be empty"
-                        ));
-                    }
-                }
-            }
-            Condition::FuzzyMatchAll(value, distance) => {
-                let default_field = default_field.ok_or_else(|| {
-                    anyhow::anyhow!(
-                        "There's no FullTextSearch field for fuzzy_match_all() function"
-                    )
-                })?;
-                if value.is_empty() {
-                    return Err(anyhow::anyhow!(
-                        "The value of fuzzy_match_all() function can't be empty"
-                    ));
-                }
-                let term = Term::from_field_text(default_field, value);
-                Box::new(FuzzyTermQuery::new(term, *distance, false))
-            }
-            Condition::All() => Box::new(AllQuery {}),
-            Condition::Or(left, right) => {
-                let left_query = left.to_tantivy_query(schema, default_field)?;
-                let right_query = right.to_tantivy_query(schema, default_field)?;
-                Box::new(BooleanQuery::union(vec![left_query, right_query]))
-            }
-            Condition::And(left, right) => {
-                let left_query = left.to_tantivy_query(schema, default_field)?;
-                let right_query = right.to_tantivy_query(schema, default_field)?;
-                Box::new(BooleanQuery::intersection(vec![left_query, right_query]))
-            }
-            Condition::Not(condition) => {
-                let query = condition.to_tantivy_query(schema, default_field)?;
-                Box::new(BooleanQuery::new(vec![
-                    (Occur::MustNot, query),
-                    (Occur::Must, Box::new(AllQuery {})),
-                ]))
-            }
-        })
-    }
-
-    /// Fields whose queries enumerate terms and therefore require full postings.
-    /// Exact positive and negative terms are exposed by `Query::query_terms`.
-    pub fn need_all_term_fields(&self) -> HashSet<String> {
-        let mut fields = HashSet::new();
-        match self {
-            Condition::StrMatch(field, ..) | Condition::Regex(field, _) => {
-                fields.insert(field.clone());
-            }
-            Condition::MatchAll(value) => {
-                if value.len() > 1 && (value.starts_with("*") || value.ends_with("*")) {
-                    fields.insert(INDEX_FIELD_NAME_FOR_ALL.to_string());
-                }
-            }
-            Condition::FuzzyMatchAll(..) => {
-                fields.insert(INDEX_FIELD_NAME_FOR_ALL.to_string());
-            }
-            Condition::Or(left, right) | Condition::And(left, right) => {
-                fields.extend(left.need_all_term_fields());
-                fields.extend(right.need_all_term_fields());
-            }
-            Condition::Not(condition) => {
-                fields.extend(condition.need_all_term_fields());
-            }
-            Condition::All()
-            | Condition::Equal(..)
-            | Condition::NotEqual(..)
-            | Condition::In(..) => {}
-        }
-        fields
-    }
-
-    #[allow(dead_code)]
-    pub fn get_tantivy_fields(&self) -> HashSet<String> {
-        let mut fields = HashSet::new();
-        match self {
-            Condition::Equal(field, _)
-            | Condition::NotEqual(field, _)
-            | Condition::In(field, ..)
-            | Condition::Regex(field, _)
-            | Condition::StrMatch(field, ..) => {
-                fields.insert(field.clone());
-            }
-            Condition::MatchAll(_) | Condition::FuzzyMatchAll(..) => {
-                fields.insert(INDEX_FIELD_NAME_FOR_ALL.to_string());
-            }
-            Condition::All() => {}
-            Condition::Or(left, right) | Condition::And(left, right) => {
-                fields.extend(left.get_tantivy_fields());
-                fields.extend(right.get_tantivy_fields());
-            }
-            Condition::Not(condition) => {
-                fields.extend(condition.get_tantivy_fields());
-            }
-        }
-        fields
-    }
-
-    // get the fields use for search in datafusion(for add filter back logical)
-    pub fn get_schema_fields(&self, fst_fields: &[String]) -> HashSet<String> {
-        let mut fields = HashSet::new();
-        match self {
-            Condition::Equal(field, _)
-            | Condition::NotEqual(field, _)
-            | Condition::StrMatch(field, ..)
-            | Condition::In(field, ..)
-            | Condition::Regex(field, _) => {
-                fields.insert(field.clone());
-            }
-            Condition::MatchAll(_) | Condition::FuzzyMatchAll(..) => {
-                fields.extend(fst_fields.iter().cloned());
-            }
-            Condition::All() => {}
-            Condition::Or(left, right) | Condition::And(left, right) => {
-                fields.extend(left.get_schema_fields(fst_fields));
-                fields.extend(right.get_schema_fields(fst_fields));
-            }
-            Condition::Not(condition) => {
-                fields.extend(condition.get_schema_fields(fst_fields));
-            }
-        }
-        fields
-    }
-
-    pub fn to_physical_expr(
-        &self,
-        schema: &arrow_schema::Schema,
-        fst_fields: &[String],
-    ) -> Result<Arc<dyn PhysicalExpr>, anyhow::Error> {
-        let cfg = get_config();
-        match self {
-            Condition::Equal(name, value) => {
-                let index = schema.index_of(name).unwrap();
-                let left = Arc::new(Column::new(name, index));
-                let field = schema.field(index);
-                let right = get_scalar_value(value, field.data_type())?;
-                Ok(Arc::new(BinaryExpr::new(left, Operator::Eq, right)))
-            }
-            Condition::NotEqual(name, value) => {
-                let index = schema.index_of(name).unwrap();
-                let left = Arc::new(Column::new(name, index));
-                let field = schema.field(index);
-                let right = get_scalar_value(value, field.data_type())?;
-                Ok(Arc::new(BinaryExpr::new(left, Operator::NotEq, right)))
-            }
-            Condition::StrMatch(name, value, case_sensitive) => {
-                create_str_match_expr(schema, name, value, *case_sensitive)
-            }
-            Condition::In(name, values, negated) => {
-                let index = schema.index_of(name).unwrap();
-                let left = Arc::new(Column::new(name, index));
-                let field = schema.field(index);
-                let values: Vec<Arc<dyn PhysicalExpr>> = values
-                    .iter()
-                    .map(|value| get_scalar_value(value, field.data_type()).map(|v| v as _))
-                    .collect::<Result<Vec<_>, _>>()?;
-                Ok(Arc::new(InListExpr::try_new(
-                    left, values, *negated, schema,
-                )?))
-            }
-            Condition::Regex(..) => {
-                unreachable!("Condition::Regex query only support for promql")
-            }
-            Condition::MatchAll(value) => {
-                let value = value
-                    .trim_start_matches("re:") // regex
-                    .trim_start_matches('*') // contains
-                    .trim_end_matches('*') // prefix or contains
-                    .to_string();
-                let term = if cfg.common.utf8_view_enabled {
-                    Arc::new(Literal::new(ScalarValue::Utf8View(Some(format!(
-                        "%{value}%"
-                    )))))
-                } else {
-                    Arc::new(Literal::new(ScalarValue::Utf8(Some(format!("%{value}%")))))
-                };
-                let mut expr_list: Vec<Arc<dyn PhysicalExpr>> =
-                    Vec::with_capacity(fst_fields.len());
-                for field in fst_fields.iter() {
-                    let term = if !cfg.common.utf8_view_enabled
-                        && let Some((_idx, schema_field)) = schema.column_with_name(field)
-                        && schema_field.data_type() == &DataType::LargeUtf8
-                    {
-                        Arc::new(Literal::new(ScalarValue::LargeUtf8(Some(format!(
-                            "%{value}%"
-                        )))))
-                    } else {
-                        term.clone()
-                    };
-                    expr_list.push(create_like_expr_with_not_null(field, term, schema));
-                }
-                if expr_list.is_empty() {
-                    return Err(anyhow::anyhow!(
-                        "Using match_all() function in a stream that don't have full text search field"
-                    )); // already check this in sql.rs
-                }
-                Ok(disjunction(expr_list))
-            }
-            Condition::FuzzyMatchAll(value, distance) => {
-                let fuzzy_expr = Arc::new(fuzzy_match_udf::FUZZY_MATCH_UDF.clone());
-                let term = if cfg.common.utf8_view_enabled {
-                    Arc::new(Literal::new(ScalarValue::Utf8View(Some(value.to_string()))))
-                } else {
-                    Arc::new(Literal::new(ScalarValue::Utf8(Some(value.to_string()))))
-                };
-                let distance = Arc::new(Literal::new(ScalarValue::Int64(Some(*distance as i64))));
-                let mut expr_list: Vec<Arc<dyn PhysicalExpr>> =
-                    Vec::with_capacity(fst_fields.len());
-                for field in fst_fields.iter() {
-                    let term = if !cfg.common.utf8_view_enabled
-                        && let Some((_idx, schema_field)) = schema.column_with_name(field)
-                        && schema_field.data_type() == &DataType::LargeUtf8
-                    {
-                        Arc::new(Literal::new(ScalarValue::LargeUtf8(Some(
-                            value.to_string(),
-                        ))))
-                    } else {
-                        term.clone()
-                    };
-                    let new_expr = Arc::new(ScalarFunctionExpr::try_new(
-                        fuzzy_expr.clone(),
-                        vec![
-                            Arc::new(Column::new(field, schema.index_of(field).unwrap())),
-                            term,
-                            distance.clone(),
-                        ],
-                        schema,
-                        Arc::new(ConfigOptions::default()),
-                    )?);
-                    expr_list.push(new_expr);
-                }
-                if expr_list.is_empty() {
-                    return Err(anyhow::anyhow!(
-                        "Using fuzzy_match_all() function in a stream that don't have full text search field"
-                    )); // already check this in sql.rs
-                }
-                Ok(disjunction(expr_list))
-            }
-            Condition::All() => Ok(Arc::new(Literal::new(ScalarValue::Boolean(Some(true))))),
-            Condition::Or(left, right) => {
-                let left = left.to_physical_expr(schema, fst_fields)?;
-                let right = right.to_physical_expr(schema, fst_fields)?;
-                Ok(Arc::new(BinaryExpr::new(left, Operator::Or, right)))
-            }
-            Condition::And(left, right) => {
-                let left = left.to_physical_expr(schema, fst_fields)?;
-                let right = right.to_physical_expr(schema, fst_fields)?;
-                Ok(Arc::new(BinaryExpr::new(left, Operator::And, right)))
-            }
-            Condition::Not(condition) => {
-                let expr = condition.to_physical_expr(schema, fst_fields)?;
-                Ok(Arc::new(NotExpr::new(expr)))
-            }
-        }
-    }
-
-    pub fn can_remove_filter(&self) -> bool {
-        match self {
-            Condition::Equal(..) => true,
-            Condition::NotEqual(..) => true,
-            Condition::StrMatch(..) => true,
-            Condition::In(..) => true,
-            Condition::Regex(..) => false,
-            Condition::MatchAll(v) => is_alphanumeric(v),
-            Condition::FuzzyMatchAll(..) => false,
-            Condition::All() => true,
-            Condition::Or(left, right) => left.can_remove_filter() && right.can_remove_filter(),
-            Condition::And(left, right) => left.can_remove_filter() && right.can_remove_filter(),
-            Condition::Not(condition) => condition.can_remove_filter(),
-        }
-    }
-}
-
-// TODO: duplication with datafusion/optimizer/physical_optimizer/utils.rs
-fn is_physical_column(expr: &Arc<dyn PhysicalExpr>) -> bool {
-    if expr.downcast_ref::<Column>().is_some() {
-        true
-    } else if let Some(expr) = expr.downcast_ref::<CastExpr>() {
-        is_physical_column(expr.expr())
-    } else {
-        false
-    }
-}
-
-// TODO: duplication with datafusion/optimizer/physical_optimizer/utils.rs
-fn get_physical_column_name(expr: &Arc<dyn PhysicalExpr>) -> &str {
-    if let Some(expr) = expr.downcast_ref::<Column>() {
-        expr.name()
-    } else if let Some(expr) = expr.downcast_ref::<CastExpr>() {
-        get_physical_column_name(expr.expr())
-    } else {
-        UNKNOWN_NAME
-    }
-}
-
-fn is_physical_value(expr: &Arc<dyn PhysicalExpr>) -> bool {
-    expr.downcast_ref::<Literal>().is_some()
-}
-
-fn get_physical_value(expr: &Arc<dyn PhysicalExpr>) -> String {
-    if let Some(literal) = expr.downcast_ref::<Literal>() {
-        match literal.value() {
-            ScalarValue::Boolean(Some(b)) => b.to_string(),
-            ScalarValue::Int64(Some(i)) => i.to_string(),
-            ScalarValue::UInt64(Some(i)) => i.to_string(),
-            ScalarValue::Float64(Some(f)) => f.to_string(),
-            ScalarValue::Utf8(Some(s)) => s.clone(),
-            ScalarValue::LargeUtf8(Some(s)) => s.clone(),
-            ScalarValue::Utf8View(Some(s)) => s.clone(),
-            ScalarValue::Binary(Some(b)) => String::from_utf8_lossy(b).to_string(),
-            _ => unimplemented!("get_physical_value not support {:?}", literal),
-        }
-    } else {
-        unreachable!()
-    }
-}
-
-// combine all exprs with OR operator
-fn disjunction(exprs: Vec<Arc<dyn PhysicalExpr>>) -> Arc<dyn PhysicalExpr> {
-    if exprs.len() == 1 {
-        exprs[0].clone()
-    } else {
-        // conjuction all expr in exprs
-        let mut expr = exprs[0].clone();
-        for e in exprs.into_iter().skip(1) {
-            expr = Arc::new(BinaryExpr::new(expr, Operator::Or, e));
-        }
-        expr
-    }
-}
-
-fn get_scalar_value(value: &str, data_type: &DataType) -> Result<Arc<Literal>, anyhow::Error> {
-    Ok(match data_type {
-        DataType::Boolean => Arc::new(Literal::new(ScalarValue::Boolean(Some(value.parse()?)))),
-        DataType::Int64 => Arc::new(Literal::new(ScalarValue::Int64(Some(value.parse()?)))),
-        DataType::UInt64 => Arc::new(Literal::new(ScalarValue::UInt64(Some(value.parse()?)))),
-        DataType::Float64 => Arc::new(Literal::new(ScalarValue::Float64(Some(value.parse()?)))),
-        DataType::Utf8 => Arc::new(Literal::new(ScalarValue::Utf8(Some(value.to_string())))),
-        DataType::LargeUtf8 => Arc::new(Literal::new(ScalarValue::LargeUtf8(Some(
-            value.to_string(),
-        )))),
-        DataType::Utf8View => {
-            Arc::new(Literal::new(ScalarValue::Utf8View(Some(value.to_string()))))
-        }
-        DataType::Binary => Arc::new(Literal::new(ScalarValue::Binary(Some(
-            value.as_bytes().to_vec(),
-        )))),
-        _ => unimplemented!(),
-    })
-}
-
-fn create_like_expr_with_not_null(
-    field: &str,
-    term: Arc<dyn PhysicalExpr>,
-    schema: &arrow_schema::Schema,
-) -> Arc<dyn PhysicalExpr> {
-    let column = Arc::new(Column::new(field, schema.index_of(field).unwrap()));
-    Arc::new(BinaryExpr::new(
-        Arc::new(IsNotNullExpr::new(column.clone())),
-        Operator::And,
-        Arc::new(LikeExpr::new(false, true, column, term.clone())),
-    ))
-}
-
-fn create_str_match_expr(
-    schema: &arrow_schema::Schema,
-    name: &str,
-    value: &str,
-    case_sensitive: bool,
-) -> Result<Arc<dyn PhysicalExpr>, anyhow::Error> {
-    let index = schema.index_of(name).unwrap();
-    let field = schema.field(index);
-    let col = Arc::new(Column::new(name, index));
-
-    // if the field is Utf8View, we need to cast it to Utf8 for str_match udf
-    let left: Arc<dyn PhysicalExpr> = if *field.data_type() == DataType::Utf8View {
-        Arc::new(CastExpr::new(col, DataType::Utf8, None))
-    } else {
-        col
-    };
-
-    // if the field is Utf8View, we need to cast it to Utf8 for str_match udf
-    let data_type = if *field.data_type() == DataType::Utf8View {
-        DataType::Utf8
-    } else {
-        field.data_type().clone()
-    };
-
-    let right = get_scalar_value(value, &data_type)?;
-    let udf = if case_sensitive {
-        Arc::new(str_match_udf::STR_MATCH_UDF.clone())
-    } else {
-        Arc::new(str_match_udf::STR_MATCH_IGNORE_CASE_UDF.clone())
-    };
-
-    let udf_expr = Arc::new(ScalarFunctionExpr::try_new(
-        udf.clone(),
-        vec![left, right],
-        schema,
-        Arc::new(ConfigOptions::default()),
-    )?);
-    Ok(udf_expr)
-}
-
-fn is_alphanumeric(s: &str) -> bool {
-    s.chars().all(|c| c.is_ascii_alphanumeric())
-}
-
-fn _is_blank_or_alphanumeric(s: &str) -> bool {
-    s.chars()
-        .all(|c| c.is_ascii_whitespace() || c.is_ascii_alphanumeric())
-}
-
-#[cfg(test)]
-mod tests {
-
-    use super::*;
-
-    #[test]
-    fn test_condition_get_tantivy_fields_equal() {
-        let condition = Condition::Equal("field1".to_string(), "value1".to_string());
-        let fields = condition.get_tantivy_fields();
-
-        assert_eq!(fields.len(), 1);
-        assert!(fields.contains("field1"));
-    }
-
-    #[test]
-    fn test_condition_get_tantivy_fields_in() {
-        let condition = Condition::In(
-            "field2".to_string(),
-            vec![
-                "value1".to_string(),
-                "value2".to_string(),
-                "value3".to_string(),
-            ],
-            false,
-        );
-        let fields = condition.get_tantivy_fields();
-
-        assert_eq!(fields.len(), 1);
-        assert!(fields.contains("field2"));
-    }
-
-    #[test]
-    fn test_condition_get_tantivy_fields_regex() {
-        let condition = Condition::Regex("field3".to_string(), "pattern.*".to_string());
-        let fields = condition.get_tantivy_fields();
-
-        assert_eq!(fields.len(), 1);
-        assert!(fields.contains("field3"));
-    }
-
-    #[test]
-    fn test_condition_get_tantivy_fields_match_all() {
-        let condition = Condition::MatchAll("search_term".to_string());
-        let fields = condition.get_tantivy_fields();
-
-        assert_eq!(fields.len(), 1);
-        assert!(fields.contains(INDEX_FIELD_NAME_FOR_ALL));
-    }
-
-    #[test]
-    fn test_condition_get_tantivy_fields_fuzzy_match_all() {
-        let condition = Condition::FuzzyMatchAll("search_term".to_string(), 2);
-        let fields = condition.get_tantivy_fields();
-
-        assert_eq!(fields.len(), 1);
-        assert!(fields.contains(INDEX_FIELD_NAME_FOR_ALL));
-    }
-
-    #[test]
-    fn test_condition_get_tantivy_fields_all() {
-        let condition = Condition::All();
-        let fields = condition.get_tantivy_fields();
-
-        assert_eq!(fields.len(), 0);
-    }
-
-    #[test]
-    fn test_condition_get_tantivy_fields_or_simple() {
-        let left = Condition::Equal("field1".to_string(), "value1".to_string());
-        let right = Condition::Equal("field2".to_string(), "value2".to_string());
-        let condition = Condition::Or(Box::new(left), Box::new(right));
-        let fields = condition.get_tantivy_fields();
-
-        assert_eq!(fields.len(), 2);
-        assert!(fields.contains("field1"));
-        assert!(fields.contains("field2"));
-    }
-
-    #[test]
-    fn test_condition_get_tantivy_fields_and_simple() {
-        let left = Condition::Equal("field1".to_string(), "value1".to_string());
-        let right = Condition::In("field2".to_string(), vec!["value1".to_string()], false);
-        let condition = Condition::And(Box::new(left), Box::new(right));
-        let fields = condition.get_tantivy_fields();
-
-        assert_eq!(fields.len(), 2);
-        assert!(fields.contains("field1"));
-        assert!(fields.contains("field2"));
-    }
-
-    #[test]
-    fn test_condition_get_tantivy_fields_or_with_overlap() {
-        let left = Condition::Equal("field1".to_string(), "value1".to_string());
-        let right = Condition::Equal("field1".to_string(), "value2".to_string());
-        let condition = Condition::Or(Box::new(left), Box::new(right));
-        let fields = condition.get_tantivy_fields();
-
-        // Should only have one field since both conditions use the same field
-        assert_eq!(fields.len(), 1);
-        assert!(fields.contains("field1"));
-    }
-
-    #[test]
-    fn test_condition_get_tantivy_fields_and_with_overlap() {
-        let left = Condition::Equal("field1".to_string(), "value1".to_string());
-        let right = Condition::Regex("field1".to_string(), "pattern.*".to_string());
-        let condition = Condition::And(Box::new(left), Box::new(right));
-        let fields = condition.get_tantivy_fields();
-
-        // Should only have one field since both conditions use the same field
-        assert_eq!(fields.len(), 1);
-        assert!(fields.contains("field1"));
-    }
-
-    #[test]
-    fn test_condition_get_tantivy_fields_nested_complex() {
-        // Create a complex nested condition: (field1 = value1 OR field2 = value2) AND (field3 =
-        // value3 OR match_all(term))
-        let left_or = Condition::Or(
-            Box::new(Condition::Equal("field1".to_string(), "value1".to_string())),
-            Box::new(Condition::Equal("field2".to_string(), "value2".to_string())),
-        );
-        let right_or = Condition::Or(
-            Box::new(Condition::Equal("field3".to_string(), "value3".to_string())),
-            Box::new(Condition::MatchAll("search_term".to_string())),
-        );
-        let condition = Condition::And(Box::new(left_or), Box::new(right_or));
-        let fields = condition.get_tantivy_fields();
-
-        assert_eq!(fields.len(), 4);
-        assert!(fields.contains("field1"));
-        assert!(fields.contains("field2"));
-        assert!(fields.contains("field3"));
-        assert!(fields.contains(INDEX_FIELD_NAME_FOR_ALL));
-    }
-
-    #[test]
-    fn test_condition_get_tantivy_fields_all_types_mixed() {
-        // Test with all different condition types mixed together
-        let equal_cond = Condition::Equal("equal_field".to_string(), "value".to_string());
-        let in_cond = Condition::In("in_field".to_string(), vec!["val1".to_string()], false);
-        let regex_cond = Condition::Regex("regex_field".to_string(), "pattern.*".to_string());
-        let match_all_cond = Condition::MatchAll("search_term".to_string());
-        let fuzzy_match_cond = Condition::FuzzyMatchAll("fuzzy_term".to_string(), 1);
-        let all_cond = Condition::All();
-
-        // Create nested structure: ((equal OR in) AND (regex OR match_all)) OR (fuzzy_match_all AND
-        // all)
-        let left_or = Condition::Or(Box::new(equal_cond), Box::new(in_cond));
-        let right_or = Condition::Or(Box::new(regex_cond), Box::new(match_all_cond));
-        let left_and = Condition::And(Box::new(left_or), Box::new(right_or));
-        let right_and = Condition::And(Box::new(fuzzy_match_cond), Box::new(all_cond));
-        let condition = Condition::Or(Box::new(left_and), Box::new(right_and));
-
-        let fields = condition.get_tantivy_fields();
-
-        assert_eq!(fields.len(), 4); // equal_field, in_field, regex_field, _all (match_all and fuzzy_match_all both use _all, so deduplicated)
-        assert!(fields.contains("equal_field"));
-        assert!(fields.contains("in_field"));
-        assert!(fields.contains("regex_field"));
-        assert!(fields.contains(INDEX_FIELD_NAME_FOR_ALL));
-    }
-
-    #[test]
-    fn test_condition_get_tantivy_fields_empty_field_names() {
-        let condition = Condition::Equal("".to_string(), "value".to_string());
-        let fields = condition.get_tantivy_fields();
-
-        assert_eq!(fields.len(), 1);
-        assert!(fields.contains(""));
-    }
-
-    #[test]
-    fn test_condition_get_tantivy_fields_special_characters() {
-        let condition = Condition::Equal("field.with.dots".to_string(), "value".to_string());
-        let fields = condition.get_tantivy_fields();
-
-        assert_eq!(fields.len(), 1);
-        assert!(fields.contains("field.with.dots"));
-    }
-
-    #[test]
-    fn test_condition_get_tantivy_fields_unicode_field_names() {
-        let condition = Condition::Equal("поле".to_string(), "значение".to_string());
-        let fields = condition.get_tantivy_fields();
-
-        assert_eq!(fields.len(), 1);
-        assert!(fields.contains("поле"));
-    }
-
-    // add some test for str_match
-    #[test]
-    fn test_str_match() {
-        let condition = Condition::StrMatch("field1".to_string(), "value1".to_string(), true);
-        let fields = condition.get_tantivy_fields();
-
-        assert_eq!(fields.len(), 1);
-        assert!(fields.contains("field1"));
-    }
-
-    #[test]
-    fn test_is_alphanumeric() {
-        assert!(is_alphanumeric("123"));
-        assert!(is_alphanumeric("123abc"));
-        assert!(!is_alphanumeric("123 abc"));
-        assert!(!is_alphanumeric("123 abc 123"));
-    }
-
-    #[test]
-    fn test_is_blank_or_alphanumeric() {
-        assert!(_is_blank_or_alphanumeric("123"));
-        assert!(_is_blank_or_alphanumeric("123abc"));
-        assert!(_is_blank_or_alphanumeric("123 abc"));
-        assert!(_is_blank_or_alphanumeric("123 abc 123"));
-    }
-
-    #[test]
-    fn test_index_condition_new() {
-        let condition = IndexCondition::new();
-        assert!(condition.conditions.is_empty());
-    }
-
-    #[test]
-    fn test_index_condition_add_condition() {
-        let mut index_condition = IndexCondition::new();
-        let condition = Condition::Equal("field1".to_string(), "value1".to_string());
-
-        index_condition.add_condition(condition.clone());
-
-        assert_eq!(index_condition.conditions.len(), 1);
-        assert!(matches!(
-            index_condition.conditions[0],
-            Condition::Equal(ref field, ref value) if field == "field1" && value == "value1"
-        ));
-    }
-
-    #[test]
-    fn test_index_condition_to_query() {
-        let mut index_condition = IndexCondition::new();
-        index_condition.add_condition(Condition::Equal("field1".to_string(), "value1".to_string()));
-        index_condition.add_condition(Condition::Equal("field2".to_string(), "value2".to_string()));
-
-        let query_string = index_condition.to_query();
-        assert_eq!(query_string, "field1=value1 AND field2=value2");
-    }
-
-    #[test]
-    fn test_index_condition_to_query_empty() {
-        let index_condition = IndexCondition::new();
-        let query_string = index_condition.to_query();
-        assert_eq!(query_string, "");
-    }
-
-    #[test]
-    fn test_index_condition_is_empty() {
-        let mut index_condition = IndexCondition::new();
-        assert!(index_condition.is_empty());
-
-        index_condition.add_condition(Condition::Equal("field1".to_string(), "value1".to_string()));
-        assert!(!index_condition.is_empty());
-    }
-
-    #[test]
-    fn test_index_condition_get_str_match_condition() {
-        let mut index_condition = IndexCondition::new();
-        index_condition.add_condition(Condition::StrMatch(
-            "field1".to_string(),
-            "value1".to_string(),
-            true,
-        ));
-
-        let result = index_condition.get_str_match_condition();
-        assert_eq!(result, Some(("value1".to_string(), true)));
-    }
-
-    #[test]
-    fn test_index_condition_get_str_match_condition_all() {
-        let mut index_condition = IndexCondition::new();
-        index_condition.add_condition(Condition::All());
-
-        let result = index_condition.get_str_match_condition();
-        assert_eq!(result, None);
-    }
-
-    #[test]
-    fn test_index_condition_is_condition_all() {
-        let mut index_condition = IndexCondition::new();
-        index_condition.add_condition(Condition::All());
-
-        assert!(index_condition.is_condition_all());
-
-        index_condition.add_condition(Condition::Equal("field1".to_string(), "value1".to_string()));
-        assert!(!index_condition.is_condition_all());
-    }
-
-    #[test]
-    fn test_condition_to_query_equal() {
-        let condition = Condition::Equal("field1".to_string(), "value1".to_string());
-        assert_eq!(condition.to_query(), "field1=value1");
-    }
-
-    #[test]
-    fn test_condition_to_query_not_equal() {
-        let condition = Condition::NotEqual("field1".to_string(), "value1".to_string());
-        assert_eq!(condition.to_query(), "field1!=value1");
-    }
-
-    #[test]
-    fn test_condition_to_query_str_match() {
-        let condition = Condition::StrMatch("field1".to_string(), "value1".to_string(), true);
-        assert_eq!(condition.to_query(), "str_match(field1, 'value1')");
-
-        let condition = Condition::StrMatch("field1".to_string(), "value1".to_string(), false);
-        assert_eq!(
-            condition.to_query(),
-            "str_match_ignore_case(field1, 'value1')"
-        );
-    }
-
-    #[test]
-    fn test_condition_to_query_in() {
-        let condition = Condition::In(
-            "field1".to_string(),
-            vec!["value1".to_string(), "value2".to_string()],
-            false,
-        );
-        assert_eq!(condition.to_query(), "field1 IN (value1,value2)");
-
-        let condition = Condition::In("field1".to_string(), vec!["value1".to_string()], true);
-        assert_eq!(condition.to_query(), "field1 NOT IN (value1)");
-    }
-
-    #[test]
-    fn test_condition_to_query_regex() {
-        let condition = Condition::Regex("field1".to_string(), "pattern.*".to_string());
-        assert_eq!(condition.to_query(), "field1=~pattern.*");
-    }
-
-    #[test]
-    fn test_condition_to_query_match_all() {
-        let condition = Condition::MatchAll("search_term".to_string());
-        assert_eq!(
-            condition.to_query(),
-            "(_all:search_term):([\"search\", \"term\"])"
-        );
-    }
-
-    #[test]
-    fn test_condition_to_query_fuzzy_match_all() {
-        let condition = Condition::FuzzyMatchAll("search_term".to_string(), 2);
-        assert_eq!(condition.to_query(), "_all:fuzzy(search_term, 2)");
-    }
-
-    #[test]
-    fn test_condition_to_query_all() {
-        let condition = Condition::All();
-        assert_eq!(condition.to_query(), "ALL");
-    }
-
-    #[test]
-    fn test_condition_to_query_or() {
-        let left = Condition::Equal("field1".to_string(), "value1".to_string());
-        let right = Condition::Equal("field2".to_string(), "value2".to_string());
-        let condition = Condition::Or(Box::new(left), Box::new(right));
-        assert_eq!(condition.to_query(), "(field1=value1 OR field2=value2)");
-    }
-
-    #[test]
-    fn test_condition_to_query_and() {
-        let left = Condition::Equal("field1".to_string(), "value1".to_string());
-        let right = Condition::Equal("field2".to_string(), "value2".to_string());
-        let condition = Condition::And(Box::new(left), Box::new(right));
-        assert_eq!(condition.to_query(), "(field1=value1 AND field2=value2)");
-    }
-
-    #[test]
-    fn test_condition_to_query_not() {
-        let inner = Condition::Equal("field1".to_string(), "value1".to_string());
-        let condition = Condition::Not(Box::new(inner));
-        assert_eq!(condition.to_query(), "NOT(field1=value1)");
-    }
-
-    #[test]
-    fn test_condition_need_all_term_fields_str_match() {
-        let condition = Condition::StrMatch("field1".to_string(), "value".to_string(), true);
-        let fields = condition.need_all_term_fields();
-        assert_eq!(fields.len(), 1);
-        assert!(fields.contains("field1"));
-    }
-
-    #[test]
-    fn test_condition_need_all_term_fields_match_all_wildcard() {
-        let condition = Condition::MatchAll("*test*".to_string());
-        let fields = condition.need_all_term_fields();
-        assert_eq!(fields.len(), 1);
-        assert!(fields.contains(INDEX_FIELD_NAME_FOR_ALL));
-    }
-
-    #[test]
-    fn test_condition_need_all_term_fields_match_all_simple() {
-        let condition = Condition::MatchAll("test".to_string());
-        let fields = condition.need_all_term_fields();
-        assert_eq!(fields.len(), 0);
-    }
-
-    #[test]
-    fn test_condition_need_all_term_fields_fuzzy_match_all() {
-        let condition = Condition::FuzzyMatchAll("test".to_string(), 2);
-        let fields = condition.need_all_term_fields();
-        assert_eq!(fields.len(), 1);
-        assert!(fields.contains(INDEX_FIELD_NAME_FOR_ALL));
-    }
-
-    #[test]
-    fn test_condition_need_all_term_fields_not_equal() {
-        let condition = Condition::NotEqual("field1".to_string(), "value".to_string());
-        let fields = condition.need_all_term_fields();
-        assert!(fields.is_empty());
-    }
-
-    #[test]
-    fn test_condition_need_all_term_fields_in_negated() {
-        let condition = Condition::In("field1".to_string(), vec!["value1".to_string()], true);
-        let fields = condition.need_all_term_fields();
-        assert!(fields.is_empty());
-    }
-
-    #[test]
-    fn test_condition_need_all_term_fields_or() {
-        let left = Condition::StrMatch("field1".to_string(), "value".to_string(), true);
-        let right = Condition::NotEqual("field2".to_string(), "value".to_string());
-        let condition = Condition::Or(Box::new(left), Box::new(right));
-        let fields = condition.need_all_term_fields();
-        assert_eq!(fields.len(), 1);
-        assert!(fields.contains("field1"));
-    }
-
-    #[test]
-    fn test_condition_need_all_term_fields_not_exact_and_expansion() {
-        let exact = Condition::Not(Box::new(Condition::Equal(
-            "field1".to_string(),
-            "value".to_string(),
-        )));
-        assert!(exact.need_all_term_fields().is_empty());
-
-        let expansion = Condition::Not(Box::new(Condition::Regex(
-            "field1".to_string(),
-            "value.*".to_string(),
-        )));
-        assert_eq!(
-            expansion.need_all_term_fields(),
-            HashSet::from(["field1".to_string()])
-        );
-    }
-
-    #[test]
-    fn test_condition_can_remove_filter_equal() {
-        let condition = Condition::Equal("field1".to_string(), "value1".to_string());
-        assert!(condition.can_remove_filter());
-    }
-
-    #[test]
-    fn test_condition_can_remove_filter_regex() {
-        let condition = Condition::Regex("field1".to_string(), "pattern.*".to_string());
-        assert!(!condition.can_remove_filter());
-    }
-
-    #[test]
-    fn test_condition_can_remove_filter_match_all_alphanumeric() {
-        let condition = Condition::MatchAll("test123".to_string());
-        assert!(condition.can_remove_filter());
-    }
-
-    #[test]
-    fn test_condition_can_remove_filter_match_all_non_alphanumeric() {
-        let condition = Condition::MatchAll("test 123".to_string());
-        assert!(!condition.can_remove_filter());
-    }
-
-    #[test]
-    fn test_condition_can_remove_filter_fuzzy_match_all() {
-        let condition = Condition::FuzzyMatchAll("test".to_string(), 2);
-        assert!(!condition.can_remove_filter());
-    }
-
-    #[test]
-    fn test_condition_can_remove_filter_or() {
-        let left = Condition::Equal("field1".to_string(), "value1".to_string());
-        let right = Condition::Equal("field2".to_string(), "value2".to_string());
-        let condition = Condition::Or(Box::new(left), Box::new(right));
-        assert!(condition.can_remove_filter());
-
-        let left = Condition::Equal("field1".to_string(), "value1".to_string());
-        let right = Condition::Regex("field2".to_string(), "pattern.*".to_string());
-        let condition = Condition::Or(Box::new(left), Box::new(right));
-        assert!(!condition.can_remove_filter());
-    }
-
-    #[test]
-    fn test_disjunction_single() {
-        use datafusion::{physical_expr::expressions::Literal, scalar::ScalarValue};
-
-        let expr = Arc::new(Literal::new(ScalarValue::Boolean(Some(true))));
-        let result = disjunction(vec![expr.clone()]);
-        assert_eq!(result.as_ref() as *const _, expr.as_ref() as *const _);
-    }
-
-    #[test]
-    fn test_get_scalar_value_boolean() {
-        let result = get_scalar_value("true", &DataType::Boolean);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_get_scalar_value_int64() {
-        let result = get_scalar_value("123", &DataType::Int64);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_get_scalar_value_utf8() {
-        let result = get_scalar_value("test", &DataType::Utf8);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_get_scalar_value_error() {
-        let result = get_scalar_value("not_a_number", &DataType::Int64);
-        assert!(result.is_err());
-    }
-
-    /// Build a minimal tantivy schema containing only the given text field names.
-    fn build_tantivy_schema(fields: &[&str]) -> Schema {
-        let mut builder = Schema::builder();
-        for name in fields {
-            builder.add_text_field(name, tantivy::schema::TEXT);
-        }
-        builder.build()
-    }
-
-    /// Collect the distinct tantivy fields referenced by a query's terms.
-    fn collect_query_fields(query: &dyn Query) -> HashSet<Field> {
-        let mut fields = HashSet::new();
-        query.query_terms(&mut |term, _| {
-            fields.insert(term.field());
-        });
-        fields
-    }
-
-    #[test]
-    fn test_to_tantivy_query_all_fields_present() {
-        // Schema has both A and B.
-        let schema = build_tantivy_schema(&["A", "B"]);
-        let field_a = schema.get_field("A").unwrap();
-        let field_b = schema.get_field("B").unwrap();
-
-        let mut cond = IndexCondition::new();
-        cond.add_condition(Condition::Equal("A".into(), "a".into()));
-        cond.add_condition(Condition::Equal("B".into(), "b".into()));
-
-        let (query, has_skipped) = cond
-            .to_tantivy_query("test", schema, None)
-            .expect("query build should succeed");
-        assert!(!has_skipped, "no field is missing, should not skip");
-
-        let referenced = collect_query_fields(query.as_ref());
-        assert!(referenced.contains(&field_a));
-        assert!(referenced.contains(&field_b));
-    }
-
-    #[test]
-    fn test_to_tantivy_query_missing_field_is_skipped() {
-        // Schema only has A; condition references B (newly added index field,
-        // no historical data in this tantivy file).
-        let schema = build_tantivy_schema(&["A"]);
-        let field_a = schema.get_field("A").unwrap();
-
-        let mut cond = IndexCondition::new();
-        cond.add_condition(Condition::Equal("A".into(), "a".into()));
-        cond.add_condition(Condition::Equal("B".into(), "b".into()));
-
-        let (query, has_skipped) = cond
-            .to_tantivy_query("test", schema, None)
-            .expect("query build should succeed even when a field is missing");
-        assert!(has_skipped, "missing field B should be reported as skipped");
-
-        let referenced = collect_query_fields(query.as_ref());
-        // Only A should be referenced; B was dropped.
-        assert_eq!(referenced.len(), 1);
-        assert!(referenced.contains(&field_a));
-    }
-
-    #[test]
-    fn test_to_tantivy_query_all_fields_missing_returns_error() {
-        // Schema has no matching fields for any condition.
-        // Should return an error so the caller falls back to parquet scan.
-        let schema = build_tantivy_schema(&["other_field"]);
-        let mut cond = IndexCondition::new();
-        cond.add_condition(Condition::Equal("A".into(), "a".into()));
-        cond.add_condition(Condition::Equal("B".into(), "b".into()));
-
-        let result = cond.to_tantivy_query("test", schema, None);
-        assert!(
-            result.is_err(),
-            "should return error when all fields are missing"
-        );
-    }
-
-    #[test]
-    fn test_to_tantivy_query_single_condition_missing_field() {
-        // Single condition whose field is missing — should return an error.
-        let schema = build_tantivy_schema(&["A"]);
-        let mut cond = IndexCondition::new();
-        cond.add_condition(Condition::Equal("B".into(), "b".into()));
-
-        let result = cond.to_tantivy_query("test", schema, None);
-        assert!(
-            result.is_err(),
-            "should return error when the only field is missing"
-        );
-    }
-}
+pub use search::index::*;
