+++ src/search/src/datafusion/distributed_plan/display.rs
