+++ src/search/src/datafusion/distributed_plan/enrich_exec.rs
