+++ src/search/src/datafusion/optimizer/physical_optimizer/enrichment.rs
@@ -24,7 +24,7 @@ use datafusion::{
     sql::TableReference,
 };
 
-use crate::service::search::datafusion::{
+use crate::datafusion::{
     distributed_plan::{
         empty_exec::NewEmptyExec, enrichment_exec::EnrichmentExec, node::RemoteScanNodes,
     },
