+++ src/search/src/datafusion/optimizer/physical_optimizer/mod.rs
@@ -13,4 +13,14 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
+#[cfg(feature = "enterprise")]
+pub mod broadcast_join;
+pub mod distribute_analyze;
+#[cfg(feature = "enterprise")]
+pub mod enrichment;
+pub mod index;
+pub mod index_optimizer;
+pub mod join_reorder;
+pub mod remote_scan;
+pub mod rewrite_match;
 pub mod utils;
